+++ packages/react/src/experimental/AiChat/index.stories.tsx
@@ -1,6 +1,5 @@
-import { CopilotKit } from "@copilotkit/react-core"
 import { Meta, StoryObj } from "@storybook/react-vite"
-import { AiChat } from "./index"
+import { AiChat, AiChatProvider } from "./index"
 
 const meta = {
   title: "Experimental/AiChat",
@@ -13,12 +12,12 @@ const meta = {
     (Story) => {
       return (
         <div className="h-full w-full bg-[hsl(0,0,98)]">
-          <CopilotKit
+          <AiChatProvider
             runtimeUrl="https://mastra.local.factorial.dev/copilotkit"
-            agent="oneAgent"
+            agent="one-workflow"
           >
             <Story />
-          </CopilotKit>
+          </AiChatProvider>
         </div>
       )
     },
@@ -30,6 +29,10 @@ type Story = StoryObj<typeof meta>
 
 export const Default: Story = {
   args: {
+    labels: {
+      greeting: "Hey Adam,",
+      initial: ["How can I help you today?"],
+    },
     children: (
       <div className="rounded border p-4">Your app content goes here</div>
     ),