+++ packages/react/src/experimental/Information/Headers/Metadata/MetadataValue.tsx
@@ -1,25 +1,22 @@
-import { Icon, IconType } from "../../../../components/Utilities/Icon"
-import { AlertCircle, Warning } from "../../../../icons/app"
-import { cn } from "../../../../lib/utils"
+import { Icon } from "@/components/Utilities/Icon"
+import { AlertCircle, Warning } from "@/icons/app"
+import { cn } from "@/lib/utils"
 import { Avatar } from "../../Avatars/Avatar"
 import { AvatarList } from "../../Avatars/AvatarList"
 import { DotTag } from "../../Tags/DotTag"
 import { RawTag } from "../../Tags/RawTag"
 import { StatusTag } from "../../Tags/StatusTag"
-import { MetadataItem, MetadataItemValue } from "./index"
+import { MetadataItem } from "./index"
 
-const DATE_ICON_STYLES: Record<
-  NonNullable<Extract<MetadataItemValue, { type: "date" }>["icon"]>,
-  { icon: IconType; iconColor: string; textColor: string }
-> = {
+const DATE_ICON_STYLES = {
   warning: {
     icon: Warning,
-    iconColor: "text-f1-icon-warning",
+    iconColor: "warning" as const,
     textColor: "text-f1-foreground-warning",
   },
   critical: {
     icon: AlertCircle,
-    iconColor: "text-f1-icon-critical",
+    iconColor: "critical" as const,
     textColor: "text-f1-foreground-critical",
   },
 }
@@ -105,9 +102,10 @@ export function MetadataValue({
       }
 
       const { icon, iconColor, textColor } = DATE_ICON_STYLES[value.icon]
+
       return (
         <div className="flex items-center justify-center gap-0.5 font-medium">
-          <Icon icon={icon} className={iconColor} />
+          <Icon icon={icon} color={iconColor} />
           <span className={textColor}>{value.formattedDate}</span>
         </div>
       )