+++ packages/react/src/experimental/OneDataCollection/grouping.ts
@@ -13,6 +13,7 @@ import { RecordType } from "./types"
 export type GroupingDefinition<R extends RecordType> = {
   /** Whether grouping is mandatory or the user can chose not to group */
   mandatory?: boolean
+  hideSelector?: boolean
   groupBy: {
     [K in RecordPaths<R>]?: {
       /** The label for the grouping */